// app/myprojects/shadowrun-in-hungary/karakterek/components/Jellemzok/Jellemzo/store/jellemzok.dto.ts
export interface JellemzokDto {
  becenev: string;
  alnev: string;
  testsuly: number;
  magassag: number;
  kedvencSzin: string;
  testalkat: string;
  borszin: string;
  szemszin: string;
  hajszin: string;
  szorszin: string;
  hajstilus: string;
  felelem: string;
  osztonzo: string;
  gyulolet: string;
  kovetendo: string;
  irtozat: string;
  vonzalom: string;
  megjelenes: string;
}